change intensity values to double from int

+++ src/main/java/com/app/fitness/fitnesprogramapp/dtos/program/CreateSetDTO.java
@@ -8,6 +8,6 @@ public class CreateSetDTO {
     private int volumeMax;
     private Long volumeMetric;
     private Long intensityMetric;
-    private int intensityMin;
-    private int intensityMax;
+    private double intensityMin;
+    private double intensityMax;
 }
